allensdk/brain_observatory/behavior/behavior_project_cache/tables/util/experiments_table_utils.py
import pandas as pd


def add_passive_flag_to_ophys_experiment_table(
        experiments_table: pd.DataFrame) -> pd.DataFrame:
    """
    adds a column to ophys_experiment_table that contains a Boolean
    indicating whether a session was passive or not based on session
    number

    Parameters
    ----------
    experiments_table: pd.DataFrame

    Returns
    -------
    experiments_table: pd.DataFrame

    Note
    ----
    Does not change the input DataFrame in-place
    """

    # Ported from
    # https://github.com/AllenInstitute/visual_behavior_analysis/blob/master
    # /visual_behavior/data_access/utilities.py#L1344

    experiments_table = experiments_table.copy(deep=True)

    experiments_table['passive'] = False

    session_25 = experiments_table.session_number.isin([2, 5])
    passive_indices = experiments_table[session_25].index.values
    experiments_table.loc[passive_indices, 'passive'] = True

    return experiments_table


def add_image_set_to_experiment_table(
        experiments_table: pd.DataFrame) -> pd.DataFrame:
    """
    Adds a column 'image_set' to the experiment_table, determined based
    on the image set listed in the session_type column string

    Parameters
    ----------
    experiments_table: pd.DataFrame

    Returns
    --------
    experiments_table: pd.DataFrame

    Notes
    -----
    Does not alter the input DataFrame in-place
    """

    # Ported from
    # https://github.com/AllenInstitute/visual_behavior_analysis/blob/master/
    # visual_behavior/data_access/utilities.py#L1403

    experiments_table = experiments_table.copy(deep=True)

    experiments_table['image_set'] = [
            session_type[15]
            if len(session_type) > 15 else 'N/A'
            for session_type
            in experiments_table.session_type.values.astype(str)]
    return experiments_table
